Update MorpheusWP.md
Lastly, Morpheus having an graphical user interface and leveraging Electron to package it as a one click install, allows Morpheus to become the first open source AI to pass the famous "Friedl test". A threshold that benchmarks when the ease of use of a software first makes it as accessible to non technical members of the general public.